+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.9.3b]
 ### Changed
 - Added channel closure in case of background reset  
+- Async reset conditions are reset every day
 
 ## [0.9.1]
 ### adapted from [v.0.7.3] version, see its changelog for more information